data: make Tense take some optional parameters
In Italian you can say "io mangio" or "mangio" and we want to accept
both here. In, say, French you have to include the personal pronoun.

Add a special case function which can do some transformations to
answers. This is a pretty heavy hammer to changing, say, “je hai” to
“j'hai”.